Fix segfault in Vt_scroll_out_all_content
The Vt_scroll_out_all_content function contains a for loop which
iterates over the visible lines on the display.  The index of this loop,
`i`, is a decrementing unsigned integer and because the condition was
that `i` be greater than or equal to the index of the top visible line,
if that line was numbered 0, the condition would never be false and `i`
would underflow.

To fix this, the loop has been changed so that `i` counts up and the
body of the loop has been changed to compensate.